♻️ Move into _lineage.py module
